pyproject.toml
[tool.poetry]
name = "snyk-project-tldr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.10"
pysnyk = "^0.6.3"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
